fix: Refactor nested calls debug tests and delete ContractDebugServiceTest
