+++ hedera-mirror-web3/src/test/java/com/hedera/mirror/web3/service/ContractCallNestedCallsHistoricalTest.java
@@ -22,8 +22,12 @@ import static com.hedera.mirror.web3.evm.utils.EvmTokenUtils.entityIdFromEvmAddr
 import static com.hedera.mirror.web3.evm.utils.EvmTokenUtils.toAddress;
 import static com.hedera.mirror.web3.utils.ContractCallTestUtil.EVM_V_34_BLOCK;
 import static com.hedera.mirror.web3.utils.ContractCallTestUtil.KEY_PROTO;
+import static com.hedera.mirror.web3.utils.OpcodeTracerUtil.OPTIONS;
+import static com.hedera.mirror.web3.utils.OpcodeTracerUtil.gasComparator;
+import static com.hedera.mirror.web3.utils.OpcodeTracerUtil.toHumanReadableMessage;
 import static com.hedera.node.app.service.evm.utils.EthSigsUtils.recoverAddressFromPubKey;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.mockito.Mockito.doAnswer;
 
 import com.google.common.collect.Range;
 import com.google.protobuf.ByteString;
@@ -33,16 +37,41 @@ import com.hedera.mirror.common.domain.token.TokenPauseStatusEnum;
 import com.hedera.mirror.common.domain.token.TokenSupplyTypeEnum;
 import com.hedera.mirror.common.domain.token.TokenTypeEnum;
 import com.hedera.mirror.common.domain.transaction.RecordFile;
+import com.hedera.mirror.web3.common.ContractCallContext;
+import com.hedera.mirror.web3.convert.BytesDecoder;
+import com.hedera.mirror.web3.evm.contracts.execution.OpcodesProcessingResult;
+import com.hedera.mirror.web3.service.model.ContractDebugParameters;
+import com.hedera.mirror.web3.utils.ContractFunctionProviderRecord;
 import com.hedera.mirror.web3.viewmodel.BlockType;
 import com.hedera.mirror.web3.web3j.generated.NestedCallsHistorical;
+import com.hedera.node.app.service.evm.contracts.execution.HederaEvmTransactionProcessingResult;
+import com.hedera.node.app.service.evm.store.models.HederaEvmAccount;
 import java.math.BigInteger;
 import java.util.Collections;
+import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
 import org.apache.tuweni.bytes.Bytes;
 import org.hyperledger.besu.datatypes.Address;
+import org.hyperledger.besu.evm.frame.ExceptionalHaltReason;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.mockito.ArgumentCaptor;
+import org.mockito.Captor;
+import org.web3j.tx.Contract;
 
+@RequiredArgsConstructor
 class ContractCallNestedCallsHistoricalTest extends AbstractContractCallServiceTest {
+
+    private final ContractDebugService contractDebugService;
+
+    @Captor
+    private ArgumentCaptor<ContractDebugParameters> paramsCaptor;
+
+    @Captor
+    private ArgumentCaptor<Long> gasCaptor;
+
+    private HederaEvmTransactionProcessingResult resultCaptor;
+    private ContractCallContext contextCaptor;
 
     private RecordFile recordFileBeforeEvm34;
 
@@ -55,6 +84,16 @@ class ContractCallNestedCallsHistoricalTest extends AbstractContractCallServiceT
         testWeb3jService.setBlockType(BlockType.of(String.valueOf(EVM_V_34_BLOCK - 1)));
         testWeb3jService.setHistoricalRange(
                 Range.closedOpen(recordFileBeforeEvm34.getConsensusStart(), recordFileBeforeEvm34.getConsensusEnd()));
+
+        doAnswer(invocation -> {
+                    final var transactionProcessingResult =
+                            (HederaEvmTransactionProcessingResult) invocation.callRealMethod();
+                    resultCaptor = transactionProcessingResult;
+                    contextCaptor = ContractCallContext.get();
+                    return transactionProcessingResult;
+                })
+                .when(processor)
+                .execute(paramsCaptor.capture(), gasCaptor.capture());
     }
 
     @Test
@@ -80,13 +119,14 @@ class ContractCallNestedCallsHistoricalTest extends AbstractContractCallServiceT
         final var contract = testWeb3jService.deploy(NestedCallsHistorical::deploy);
 
         // When
-        final var result =
-                contract.call_nestedGetTokenInfo(tokenAddress.toHexString()).send();
+        final var function = contract.call_nestedGetTokenInfo(tokenAddress.toHexString());
+        final var result = function.send();
         // Then
         assertThat(result).isNotNull();
         assertThat(result.token).isNotNull();
         assertThat(result.deleted).isFalse();
         assertThat(result.token.memo).isEqualTo(tokenMemo);
+        verifyOpcodeTracerCall(function.encodeFunctionCall(), contract);
     }
 
     @Test
@@ -111,8 +151,8 @@ class ContractCallNestedCallsHistoricalTest extends AbstractContractCallServiceT
         final var contract = testWeb3jService.deploy(NestedCallsHistorical::deploy);
 
         // When
-        final var result = contract.call_nestedHtsGetApproved(tokenAddress.toHexString(), BigInteger.ONE)
-                .send();
+        final var function = contract.call_nestedHtsGetApproved(tokenAddress.toHexString(), BigInteger.ONE);
+        final var result = function.send();
 
         // Then
         final var key = ByteString.fromHex(spenderPublicKey);
@@ -120,6 +160,7 @@ class ContractCallNestedCallsHistoricalTest extends AbstractContractCallServiceT
                         Bytes.wrap(recoverAddressFromPubKey(key.substring(2).toByteArray())))
                 .toString();
         assertThat(result).isEqualTo(expectedOutput);
+        verifyOpcodeTracerCall(function.encodeFunctionCall(), contract);
     }
 
     @Test
@@ -154,16 +195,16 @@ class ContractCallNestedCallsHistoricalTest extends AbstractContractCallServiceT
         final var contract = testWeb3jService.deploy(NestedCallsHistorical::deploy);
 
         // When
-        final var result = contract.call_nestedMintToken(
-                        tokenAddress.toHexString(),
-                        BigInteger.ZERO,
-                        Collections.singletonList(
-                                ByteString.copyFromUtf8("firstMeta").toByteArray()))
-                .send();
+        final var function = contract.call_nestedMintToken(
+                tokenAddress.toHexString(),
+                BigInteger.ZERO,
+                Collections.singletonList(ByteString.copyFromUtf8("firstMeta").toByteArray()));
+        final var result = function.send();
 
         // Then
         int expectedTotalSupply = nftAmountToMint + 1;
         assertThat(result).isEqualTo(BigInteger.valueOf(expectedTotalSupply));
+        verifyOpcodeTracerCall(function.encodeFunctionCall(), contract);
     }
 
     private EntityId nftPersistHistorical(
@@ -273,5 +314,66 @@ class ContractCallNestedCallsHistoricalTest extends AbstractContractCallServiceT
                                 recordFileBeforeEvm34.getConsensusStart(), recordFileBeforeEvm34.getConsensusEnd())))
                 .persist();
         return spenderEntityId;
+    }
+
+    private void verifyOpcodeTracerCall(final String callData, final Contract contract) {
+        ContractFunctionProviderRecord functionProvider = ContractFunctionProviderRecord.builder()
+                .contractAddress(Address.fromHexString(contract.getContractAddress()))
+                .build();
+
+        final var callDataBytes = Bytes.fromHexString(callData);
+        final var debugParameters = ContractDebugParameters.builder()
+                .block(functionProvider.block())
+                .callData(callDataBytes)
+                .consensusTimestamp(domainBuilder.timestamp())
+                .gas(15_000_000L)
+                .receiver(functionProvider.contractAddress())
+                .sender(new HederaEvmAccount(functionProvider.sender()))
+                .value(functionProvider.value())
+                .build();
+
+        if (functionProvider.expectedErrorMessage() != null) {
+            verifyThrowingOpcodeTracerCall(debugParameters, functionProvider);
+        } else {
+            verifySuccessfulOpcodeTracerCall(debugParameters);
+        }
+        assertThat(paramsCaptor.getValue()).isEqualTo(debugParameters);
+        assertThat(gasCaptor.getValue()).isEqualTo(debugParameters.getGas());
+    }
+
+    @SneakyThrows
+    private void verifyThrowingOpcodeTracerCall(
+            final ContractDebugParameters params, final ContractFunctionProviderRecord function) {
+        final var actual = contractDebugService.processOpcodeCall(params, OPTIONS);
+        assertThat(actual.transactionProcessingResult().isSuccessful()).isFalse();
+        assertThat(actual.transactionProcessingResult().getOutput()).isEqualTo(Bytes.EMPTY);
+        assertThat(actual.transactionProcessingResult())
+                .satisfiesAnyOf(
+                        result -> assertThat(result.getRevertReason())
+                                .isPresent()
+                                .map(BytesDecoder::maybeDecodeSolidityErrorStringToReadableMessage)
+                                .hasValue(function.expectedErrorMessage()),
+                        result -> assertThat(result.getHaltReason())
+                                .isPresent()
+                                .map(ExceptionalHaltReason::getDescription)
+                                .hasValue(function.expectedErrorMessage()));
+        assertThat(actual.opcodes().size()).isNotZero();
+        assertThat(toHumanReadableMessage(actual.opcodes().getLast().reason()))
+                .isEqualTo(function.expectedErrorMessage());
+    }
+
+    private void verifySuccessfulOpcodeTracerCall(final ContractDebugParameters params) {
+        final var actual = contractDebugService.processOpcodeCall(params, OPTIONS);
+        final var expected = new OpcodesProcessingResult(resultCaptor, contextCaptor.getOpcodes());
+        // Compare transaction processing result
+        assertThat(actual.transactionProcessingResult())
+                .usingRecursiveComparison()
+                .ignoringFields("logs")
+                .isEqualTo(expected.transactionProcessingResult());
+        // Compare opcodes with gas tolerance
+        assertThat(actual.opcodes())
+                .usingRecursiveComparison()
+                .withComparatorForFields(gasComparator(), "gas")
+                .isEqualTo(expected.opcodes());
     }
 }
